+++ packages/react-components-lab/src/components/Syntax/types.ts
@@ -1,0 +1,9 @@
+type Language = 'typescript';
+
+interface SyntaxProps {
+  code?: string;
+  language?: string;
+}
+
+export { SyntaxProps, Language };
+export default SyntaxProps;