+++ packages/react-components-lab/src/index.ts
@@ -4,6 +4,21 @@ export * from './components/Legend/Legend';
 export { default as LegendStyles } from './components/Legend/styles';
 export * from './components/Legend/types';
 
+// BarLegend
+export { default as BarLegend } from './components/BarLegend/BarLegend';
+export { default as BarLegendStyles } from './components/BarLegend/styles';
+export * from './components/BarLegend/types';
+
+// Loader
+export { default as Loader } from './components/Loader/Loader';
+export { default as LoaderStyles } from './components/Loader/styles';
+export * from './components/Loader/types';
+
+// GapBox
+export { default as GapBox } from './components/GapBox/GapBox';
+export { default as GapBoxStyles } from './components/GapBox/styles';
+export * from './components/GapBox/types';
+
 // Snackbar
 export { default as SnackbarProvider } from './components/Snackbar/SnackbarProvider';
 export { default as useSnackbar } from './components/Snackbar/useSnackbar';
@@ -13,3 +28,8 @@ export * from './components/Snackbar/types';
 // ThemeProvider
 export { default as ThemeProvider } from './components/ThemeProvider/ThemeProvider';
 export * from './components/ThemeProvider/types';
+
+// Hooks
+export { default as useBelow } from './hooks/useBelow';
+export { default as useLocalStorage } from './hooks/useLocalStorage/useLocalStorage';
+export * from './hooks/useLocalStorage/types';